Return message if the machine has insufficient change
Simple code to meet base behaviour expectation. If the change supply is
less than the value of change needed to reimburse the customer, the
transaction cannot continue. Obviously this is the very base form of
this logic, and in reality the change supply would be supplemented by
the payment first, and in the case of this feature spec, would be able
to provide change. However, in reality, the vending machine is going to
attempt to give the greatest amount of the largest value of coin it has
first, and then move on to the next largest value of coin etc. until the
value of change is met. This functionality is more complex and can come
later.