Introduce ChangeHandler
Delegates the responsiblity of calculating the change after payment to
the ChangeHandler. The ChangeHandler now holds the state of the
machine's internal change supply. This method of calculating change is
in it's simplest form. The next step would be to incorporate logic that
takes into consideration individual coins and their quantity.